[TASK] Update phpstan/phpstan-phpunit to 2.0.8
PHPStan contains a bug [1] where ordering of intersections matter
for method calls, where the first matching method of an intersection
wins since 1.12.8.

This error was already exposed in our v12.4 branch some
time ago (and was fixed via #105559), but is now also
happending in main and 13.4 with the update to phpstan-phpunit
v2.0.8 (because phpstan-phpunit calls `ObjectType::isSuperTypeOf(…)`
which has side effects and changes intersection orderings,
that are invoked when phpstan-phpunit uses this utility method).

Since the bug is confirmed upstream we apply the same strategy as for
our 12.4 branch and add these errors as false positives to our baseline
for the time being.

Commands executed:
  composer require --dev phpstan/phpstan-phpunit:^2.0.8 -W
  Build/Scripts/runTests.sh -s phpstanGenerateBaseline

[1] https://github.com/phpstan/phpstan/issues/11978

Resolves: #108107
Releases: main, 13.4
Change-Id: I4f81ef3a4f67987cedf58a3cff9075874d607f0e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91630
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>