[!!!][TASK] Avoid global content link rewriting in FE
The frontend rendering contains a method that receives
the fully generated Response body content to apply a
global search and replace string operation in order
to fix links to included assets.

This was always a bad idea and is now finally obsolete
with the new system resource API: The system can expect
that links to resources are final when they are created
by some sub system.

The patch removes the logic along with a TYPO3_CONF_VARS
setting that allowed exending this brittle magic.

Resolves: #108114
Related: #107537
Releases: main
Change-Id: Ib67aad945475a93e774338c621df6a5641e5f36d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91656
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Lina Wolf <112@linawolf.de>
Reviewed-by: Helmut Hummel <typo3@helhum.io>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Lina Wolf <112@linawolf.de>
Tested-by: Helmut Hummel <typo3@helhum.io>