[BUGFIX] Ensure f:*.action VHs use current page if pageUid is empty
Change #107956 accidentally changed behavior of the `f:*.action`
viewhelpers in terms of pageUid handling within Extbase context.
Before the change an empty pageUid attribute (`""`) would cause
to fall back to the current page.

After the change this was treated as invalid page,
causing the whole `a` tag to vanish.

This patch restores the former behavior and documents this with a test.

Resolves: #108164
Related: #107956
Releases: main, 13.4
Change-Id: I7d1d7b0382e3fe73c0e3e70b9107fc8811f2e48b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91712
Tested-by: core-ci <typo3@b13.com>
Tested-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Moritz Ngo <moritz.ngo@kandoh.de>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Moritz Ngo <moritz.ngo@kandoh.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>