[TASK] Rename “Site Management” to “Sites” and submodule “Sites” to “Setup”
The top-level module "Site Management" is renamed
to "Sites" to improve clarity in the backend. The
new name more accurately reflects the purpose of
the module as an overview of all configured websites
and aligns better with TYPO3’s concise module
naming strategy.

Additionally, the submodule previously also named
"Sites" is therefore renamed to "Setup". This avoids
name duplication and provides a clearer description
of its role as the central configuration interface
for setting up sites (languages, domains, and routing
behavior).

Resolves: #108118
Related: #107628
Releases: main
Change-Id: I374fbc120497ae274a84cf6f8d7c35b55a8ca03c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91795
Tested-by: Lina Wolf <112@linawolf.de>
Tested-by: Oli Bartsch <bo@cedev.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Lina Wolf <112@linawolf.de>
Reviewed-by: Oli Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>