[TASK] Disable db analyzer ac test in composer mode
Patch for issue #108246 fixed an issue that the composer
mode acceptance tests always used sqlite and not mysql
or others as configured.

This revealed an issue found by nightlies that tables
created by 'bin/typo3 setup' with mysql (mariadb?) are
not set up with proper collation. The install tool DB
analyzer acceptance test finds this.

We were unable to come up with a quick solution to this
and created #108257 for a deeper look at the issue. Due
to current time constraints with version 14.0 being close
by, and green nightly having priority, the patch disables
the failing test in composer mode for now.

Resolves: #108256
Related: #108246
Related: #103297
Related: #105540
Releases: main
Change-Id: I3eebafab6fd3e0871b0f085eeaeccb3c05b86d3a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91812
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Jasmina Ließmann <typo3@minapok.eu>
Reviewed-by: Jasmina Ließmann <typo3@minapok.eu>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Sven Liebert <mail@sven-liebert.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Sven Liebert <mail@sven-liebert.de>