[TASK] runTests.sh: Add command to run GUI in the browser
Currently, the option "-s e2e-prepare" prepares a
composer based instance of TYPO3 to run the
end-to-end tests on and shows commands to run
the test using the GUI. But this requires a local
installation of node/npm.

So the option "-s e2e-browser" has been added to
run the Playwright GUI inside a container and make
it available in the browser on http://127.0.0.1:<RANDOM_PORT>.
This way additional local dependencies are unnecessary.

Resolves: #108266
Releases: main
Change-Id: I8c920962f482cfcd8952d8b430d632b291462abe
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91824
Tested-by: Sven Liebert <mail@sven-liebert.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Sven Liebert <mail@sven-liebert.de>