[TASK] Always upload Playwright reports to GitLab
Currently, Playwright reports are only uploaded
to GitLab on-failure. This means there is no report
to download for flaky tests.

This is now changed to always upload the reports no
matter what the job status is. These artifacts expire
after 1 week, to keep storage usage to a minimum.

Resolves: #108268
Releases: main
Change-Id: I03ae0cc2bcd6473667511af916f2e158de97740d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91826
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>