[BUGFIX] Boot non-failsafe state in `cache:flushtags` command
With #104054 the `cache:flushtags` command has been introduced
to allow flushing cache(s) based on one or more tags.

The command is registered in `EXT:core` using `ServiceProvider`,
making the command available in failsafe context simply using
the cache manager, which has most of the caches configured as
`NullBackend` (aka install tool context) and not the real cache
configuration, which means that tags are not really removed.

This change modifes the command to recieve the bootservice
instead of the cache manager, booting up the non-failesafe
mode and grapping the CacheManager from non-failsafe mode,
having the real configuration for all caches in place.

To cover what has been missed with the original implementation
the `CacheFlushTagsCommandTest` functional test is added, which
already states the broken state before the included change and
demonstrating it is now possible to remove tags correctly for
example from the page cache.

Resolves: #108280
Related: #104054
Releases: main
Change-Id: Ib703d20d877ba0bf869cd551e2d046f6d34b527f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91839
Tested-by: Sven Liebert <mail@sven-liebert.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Sven Liebert <mail@sven-liebert.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>