[BUGFIX] Fix wrong language key migrations
The list of the field migrations was wrong as the mentioned migrations
didn't happen that way. Instead, the showitems were migrated in a way
that the original column label is used from now on with very few
exceptions where no column exists like 'hidden' for example.

Resolves: #108302
Related: #107900
Related: #107913
Releases: main
Change-Id: I7b9089ab4c3313b97d66c91387d05a20dd13c0b7
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91865
Tested-by: Lina Wolf <112@linawolf.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Lina Wolf <112@linawolf.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>