[!!!][TASK] Use composer.json only for extension title
In composer.json there is no "title" field, only "description"
However since #96228 the description fields in all core extensions
have the extension title incorporated into the composer.json
description field, by separating the title from the description
with a dash " - "

This is now used to get rid of reading the ext_emconf.php
in Composer mode entirely, by exploding the
composer.json description field by " - " and then
using the part before this sequence as title and the second
part as description. If the sequence is not found the entire
description will also be used as title.

So the impact of this change, while still technically breaking,
is pretty low and will only affect how extension titles are shown
in extension manager or extension:list command.

Resolves: #108304
Related: #96228
Releases: main
Change-Id: I9520cc2c78b801ab6833d6047b5a7bddb7980580
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91874
Tested-by: Helmut Hummel <typo3@helhum.io>
Reviewed-by: Helmut Hummel <typo3@helhum.io>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Lina Wolf <112@linawolf.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Lina Wolf <112@linawolf.de>