[FEATURE] Allow number of decimals in `stdWrap.bytes` function
The TypoScript function `stdWrap.bytes` now accepts an additional
configuration parameter `decimals`. It allows to explicitly define the
number of decimals in the resulting number representation. By default,
the number of decimals is derived from the formatted size.

In addition, the consumed PHP function `GU::formatSize()` is extended
as well. The additional parameter `$decimals` is added and defaults to
`null`, which results in the same behavior as for the TypoScript
function `stdWrap.bytes`.

Resolves: #108344
Releases: main
Change-Id: I029c67c34cc6935e7736b114104be185f1bf8b6c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91907
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>