[BUGFIX] Show select icons for TCA folder config
It is possible to configure a select box to pull
icons from a folder in an extension. To do so,
the folder contents needs to be pulled in, but
for icons to be shown not the absolute file system
path must be provided to the icon API, but a
resource identifier.

For now a temporary fix is applied, so that the icons
are show again. To properly fix this without calling
internal API, the system resource API needs to be extended
with folder objects, with which resource identifiers from
within this folder can be pulled in.

Resolves: #108367
Related: #107537
Releases: main
Change-Id: I170d228a93f82051fcdb72a7f057f6d049e67450
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91934
Tested-by: Helmut Hummel <typo3@helhum.io>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Helmut Hummel <typo3@helhum.io>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>