[BUGFIX] Harden resource identifier recognition
Instead of checking for string prefix,
a pre-parsing is now applied to the resource
identifier, by exploding by : and using the first part
as resource type.

By doing so, the multiple if expression can now be exchanged
by a single match expression for extra clarity.

Tests are added to prove additional robustness.

Resolves: #108368
Related: #107537
Releases: main
Change-Id: I4ae6d5e0a33e036dcb4d7019a4bd62f0a64ef363
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91938
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Helmut Hummel <typo3@helhum.io>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Helmut Hummel <typo3@helhum.io>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Garvin Hicking <garvin@hick.ing>