[TASK] Disable shared memory usage for chrome testing images
Selenium suggests [1] to use a container shared memory size of 2GB
(to be passed via `--shm-size 2g` to docker) because chrome needs
more shared memory than provided by default in docker or podman.

It seems that the `tmpfs` mount on `/dev/shm` has been applied in #93575
(which "to deal with changed exceptions and behaviour"), for the same
reason.

runTests.sh executions with current docker versions (instead of our
default podman environment) revealed that the tmpfs mount leads
chrome into crashing from time to time.

Our testing infrastructure does not allow to specify `--shm-size 2g`
because our test runners do not run in host IPC namespace, but there
is another, better solution nowadays.

In the meantime chrome added support for disabling /dev/shm usage and
the developer suggested to use that:
https://issues.chromium.org/issues/40527382#comment77

For that reason we disable the tmpfs workaround and use the officially
supported cmdline flag instead.

[1] https://hub.docker.com/r/selenium/standalone-chrome
> When executing docker run for an image that contains a browser please
> use the flag --shm-size=2g to use the host's shared memory.

Releases: main, 13.4, 12.4
Resolves: #108380
Related: #93575
Change-Id: I77ac7b07dc01df54ed0edfa56c15baa1c0c620ad
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/92553
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>