[TASK] Make use of proper formated XLIFF files more robust
If an XLIFF file is a translation, then `target` items must be used.
If an XLIFF file is a translation base e.g. 'en', then `source` items
must be used.

This decision is only dependant on the file itself and is not dependant
on the 'locale' context in which a file is loaded.

E.g. in Localization 'en' files might be added to a non-english locale.
Then Symfony Translator calls XliffLoader with locale like 'ru' and an
english XLIFF file.

With this change XLIFF files which have a target language attribute

* `target-language` (XLIFF 1.2)
* `trgLang`(XLIFF 2.0)

are handled as localized file and all values are extracted from
`target` tags.
XLIFF files which do not have a target language attribute are
handled as original file and all values are extracted from
`source` tags.
There are no fallbacks in place.

It was a feature - and became a bug!

On all these changes which are not directly related to `XliffLoader`:
The implementation was buggy until now. Every language file had its
english version as fallback added.

Example for locale 'fr-FR':

* fr-FR -> en
* fr -> en
* en -> en

As a result if there was no language file 'fr-FR' all labels used
fallback 'en'.
Instead they should first fallback to 'fr' - and use 'en' as final
fallback only:

* fr-FR -> fr -> en

So why was this behaviour not buggy until now?

For all files loaded in the context of non 'en' locales, the
`XliffLoader` implemented strict usage of `target` items only, ignoring
all `source` items.

1. There is no 'fr-FR' file.
2. The 'en' file is loaded as fallback.
3. `XliffLoader` extracts all `target` items from 'en' files. As there
   are only `source` items in 'en' files this results in no items
   found, so every item is still `null`.
4. The 'fr' file is loaded as fallback and all `target` items are
   applied.
5. The 'en' file is loaded (this time in 'en' locale context) and all
   `source` items are applied as fallback.

So not having any `target` items in 'en' files was used as feature
until now.

How to get on from here?

The localization fallback and override should be based on
Symfony Translator `fallbackCatalogue`.

Resolves: #108386
Related: #107436
Releases: main
Change-Id: I78bf3166c51f78244c7125f77a82e5eb9fe6a66c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91994
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>