[BUGFIX] Avoid caching content twice in cache_pages
With TypoScriptFrontendController being gone we can start
looking at deeper constructs that surface their issues now.

The PageRenderer is particularly delicate. It becomes more
and more obvious this construct needs an overhaul. We will
try to isolate it to open up development.

As a starter, we'll fix a rather nasty issue: In frontend,
when non-cached elements ("INT") are involved, the generated
content containing cached elements and placeholders for the
non-cached elements is cached twice in cache_pages: Once as
'content' array key, and a second time as PageRenderer state
via $bodyContent property. With "fully cached" render() is
called that triggers reset() at the end to clear state. With
non-cached elements involved, renderPageWithUncachedObjects()
does not reset state, so $bodyContent is later part of the
cached PageRenderer state. There are various options to
approach this with one hindrance being the render-postProcess
hook, which receives $bodyContent and should continue to
receive it to not break existing consumers.

The solution is to for now reset $bodyContent in
renderPageWithUncachedObjects() to have the empty string
cached instead, and then reset this property just before
the hook is called. This solution catches both "empty cache"
and "got cached page, calculate non-cached elements now"
scenarios. The patch also inlines two protected methods to
make the main methods a bit more obvious.

The patch is not an answer to the overall situation around
"full" and "partial" content caching. The responsibilities
who takes care of it are not sorted out and there is a happy
mixture of this state being set and hand over. We'll have to
work on that with future patches.

Note, while "the main content is cached twice" sounds really
bad, it isn't in practice: cache_pages is usually gzip'ed,
which suppresses long duplicate strings effectively. This is
why we don't work on a v13 backport of this bugfix: Its ugly,
but the impact is limited and it would be much harder to fix
in v13 with the many refactorings that happened around TSFE
in v14.

Resolves: #108388
Releases: main
Change-Id: I52c90cccb0c9ab61db6e509dc771d0d87d09e748
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91996
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>