[TASK] Playwright: Limit workers when running locally
Currently, Playwright uses the available
CPUs / 2 to determine the number of
workers to use. This means when enough CPUs are
available all tests run at the same time which
causes the php-fpm to bail out.

This has been fixed by limiting the max
amount of workers to 4. So max 4 tests run
at the same time.

Resolves: #108392
Releases: main
Change-Id: Ibdf4f19a4913079de12604facafd13079a22e905
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91998
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Benni Mack <benni@typo3.org>