[BUGFIX] Relax `QueryObjectModelFactory::statement()`
It's a lived practice to pass QueryBuilder instances to the
extbase `Query->statement()` within extbase repositories to
allow execution of custom queries.

Internally this is passed to the internal factory class
`\TYPO3\CMS\Extbase\Persistence\Generic\Qom\QueryObjectModelFactory`
also allowing the three types for the statement which uses
`GeneralUtility::makeInstance()` to create a instance of
`\TYPO3\CMS\Extbase\Persistence\Generic\Qom\Statement`.

`Query->statement()` and `Statement::__construct()` allows
passing following types as statement:

* string as $statement
* \TYPO3\CMS\Core\Database\Query\QueryBuilder as $statement
* \Doctrine\DBAL\Statement as $statement

In the past passing a QueryBuilder instance through the chain
worked, albeit `QueryObjectModelFactory::statement()` had only
`string` stated as PHP-Docblock and have been replaced with
native type `string` in #106416 breaking the chain now silently
converting the QueryBuilder to a string and pasing it down.

Looking not that much of an issue, this leads to a different code
path execution in `Typo3DbBackend::getObjectDataByQuery()` and in
some cases to different results as the different results.

This change adopts the additional types in the method signature of
`QueryObjectModelFactory::statement()` to mitigate converting to
string and changing the behaviour. A confusing todo comment in
`Typo3DbBackend->getObjectDataByQuery()` is also removed because
`Qom\Statement->getStatement()` only returning itself a Qom\Statement
does not make much sense or need to have the QueryBuilder conversion
moved to it first.

Resolves: #108402
Related: #106416
Releases: main
Change-Id: I5037af9720ab828bd840fe564920f75a968fb2a8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/92004
Reviewed-by: Sebastian Fischer <typo3@evoweb.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Torben Hansen <derhansen@gmail.com>