[BUGFIX] Ensure EXT:scheduler records show interval frequency
The new TCA-based EXT:scheduler allows to set a task's frequency
with a "cronCmd" ("0 9,15 * * 1-5", ...) or an "interval" (seconds).

Saving a task with the "internval" configuration through the combobox
persisted the right data, but when re-displaying the value via
FormEngine, the persisted data would always only show the contents
of "cronCmd" (empty).

With this patch, in case "cronCmd" is empty, the interval will be
evaluated and used for the contents of the combobox.

This follows the same approach as in line 283 of
`EXT:scheduler/Classes/Domain/Repository/SchedulerTaskRepository.php`:

```
$taskData['frequency'] = $taskObject->getExecution()->getCronCmd()
                       ?: $taskObject->getExecution()->getInterval();
```

Resolves: #108398
Related: #107488
Releases: main, 14.0
Change-Id: I9f866d5edeb2f1f5226a9add2ba8457fc3db5654
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/92012
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Lina Wolf <112@linawolf.de>
Reviewed-by: Oli Bartsch <bo@cedev.de>
Tested-by: Wolfgang Wagner <wwagner@wwagner.net>
Tested-by: Lina Wolf <112@linawolf.de>
Tested-by: Oli Bartsch <bo@cedev.de>