[FEATURE] Add "datetimesec" TCA FormEngine format option
Currently, TCA only supports flatpickr date instances
with a "HH:MM" picker to enter the hour and minute, but
not "HH:MM:SS" to also support adding seconds.

There is only a dedicated "timesec" format, which does
not have a date associated with it.

This patch adds a "datetimesec" option, that acts just
like "datetime" would, but also support entering
seconds. This format is only valid for the integer-based
storage `dbType`, or `dbType=datetime`. Using other
`dbType` settings ("date", "time") will throw an
exception hinting at misconfiguration.

Tests are adapted to cover the scenario with additional
second input.

With a drive-by change a phpstan baseline entry is
removed that covers a misaligned phpdoc adjustment
so that $itemValue could also carry a NULL value.

Resolves: #108431
Releases: main
Change-Id: Ief616741d3502b481cc5f62614952171f875692a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/92056
Tested-by: Josua Vogel <j.vogel97@web.de>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Tested-by: Garvin Hicking <garvin@hick.ing>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Josua Vogel <j.vogel97@web.de>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>