[TASK] Adjust and relocate runTests watchRst templates
The .rst template files are now moved into a distinct
directory to unclutter the `Scripts` subdirectory.

Also, the rst templates now contain a default example index
"Backend" to create a valid rst file for generation.

Third, a typo is corrected.

Last and least, the runTests help is enhanced with a
few examples.

Resolves: #108459
Related: #108219
Related: #108066
Releases: main, 13.4, 12.4
Change-Id: Idb9ef187f7d5dc6c6cce11022b4295d52ed47444
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/92102
Tested-by: core-ci <typo3@b13.com>
Tested-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Garvin Hicking <garvin@hick.ing>