[FEATURE] Allow to configure "first day of week" datepicker setting
Previously, the date-time picker used the selected locale of a
backend user.

However, certain locale configurations might rather be a user
preference. For example, users may prefer an english backend
but want to use a weekday start on "Monday" instead of
"Sunday" due to their cultural habits.

Thus, the "first day of a week" has been decoupled from the
locale selection and can be configured on a per-user setting.
By default, it still inherits the locale's default' setting,
if not changed (for example, english=sunday and german=monday).

Inside the user settings and tab panel `Backend appearance`,
a new dropdown `First day of week (Date picker)` appears.

The setting is stored in both the persisted `be_users.uc`
preference blob, as well as on the JavaScript-persistence side.

Sadly PHP's ICU does not allow locale-specific iteration
of all available week days, and we don't require a package like
symfony/intl, so an algorithm had to be implemented to create
timestamps for all 7 days to read their locale-specific names
and offer as a dropdown in the fake TCA user settings.

Resolves: #108463
Releases: main
Change-Id: Ib1c22c61cb70536502c4cc3a49a685ee9b7d4d24
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/92083
Tested-by: Josef Glatz <typo3@josefglatz.at>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Guido Schmechel <guido.schmechel@brandung.de>
Reviewed-by: Guido Schmechel <guido.schmechel@brandung.de>
Tested-by: Garvin Hicking <garvin@hick.ing>