[BUGFIX] Properly place TCA colorpicker in HTML dialogs
With the introduction of HTML native dialog-modals in the backend,
the placement TCA-type "Color" broke, as the color-picker will be
placed outside the `<dialog>`-element. [1]

To solve this issue, the parent-option of the "Alwan"-color-picker
is now defined point to the nearest dialog element. This
option has been added with alwan 2.3.0. [2]

[1] https://github.com/sefianecho/alwan/issues/17
[2] https://github.com/sefianecho/alwan/releases/tag/v2.3.0

Resolves: #108488
Related: #107443
Releases: main
Change-Id: Idcc8e3921e89bce5b7db6351488a3bca0817c3a8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/92124
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Tested-by: core-ci <typo3@b13.com>