[TASK] Clean up EditDocumentController
This change cleans up some logic around persisting
records when editing in FormEngine.

It also inlines the methods init(), and preInit() and
main() into mainAction() so the flow of the execution
is more obvious, without obfuscated ResponseInterface
parts.

Superfluous protected properties have been removed.

This is a first step to split up processing
and rendering of records within FormEngine in the
future, and deal with real redirect() and forward()
logic, as well as clarifying the responsibilities.

Resolves: #108525
Releases: main
Change-Id: I80e617232cbcf3c7598b39779538127a0a2b4195
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/92169
Tested-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: core-ci <typo3@b13.com>