[TASK] Add typehint int for method doesPageHaveUnallowedTables
This method is only used in one place in the checkValue method.
There the call to getDefaultLanguagePageId already ensures that
the $recordId is an int.

The check for "not a number" doesn't make any sense anymore.
For safety measures, we can keep an explicit check for zero.
However, it's very unlikely to be zero. The only imaginable
scenario could be userland code providing bogus parameters.

Resolves: #108552
Releases: main
Change-Id: I1b758ed410d746e8e5d5b8edd4b4e1aaa6f88c8f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/92198
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>