[TASK] Retrieve available Page Types from TCA sub schemata
The method PageDoktypeRegistry->getRegisteredDoktypes is not really
needed. We have all information about available Page Types in the
TcaSchema. The "NewMultiplePages" view was the only consumer of this
method, which is now migrated to use TcaSchema instead.

In the end, the "Registry" part is a bit misleading. The functionality
is rather to add additional restrictions which records may be created
on that Page Type. This is totally optional, if the default restriction
is good enough.

Resolves: #108556
Releases: main
Change-Id: I2c574ffcbc060abf839fb9ee3afda1e8564be5c1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/92202
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Georg Ringer <georg.ringer@gmail.com>